Move unit tests to proper unit folder
